BuildRequest function separated and accpects api as argument, instead of obejctStorageAPI.
